+++ version_test.go
@@ -8,7 +8,11 @@ import (
 func TestVersions(t *testing.T) {
 	ocvv := OpenCVVersion()
 
-	if !(strings.Contains(ocvv, "4.11") || strings.Contains(ocvv, "4.12") || strings.Contains(ocvv, "4.13")) {
+	// TODO CV5: we should probably remove the 4.x checks
+	if !(strings.Contains(ocvv, "4.11") ||
+		strings.Contains(ocvv, "4.12") ||
+		strings.Contains(ocvv, "4.13") ||
+		strings.HasPrefix(ocvv, "5.")) {
 		t.Error("Wrong version of OpenCV:", ocvv)
 	}